fix: 리뷰 평점 모델 변경

+++ review/models.py
@@ -13,7 +13,7 @@ class Review(models.Model):
     order = models.OneToOneField(Order, on_delete=models.SET_NULL, related_name="review", null=True)
     caption = models.CharField(max_length=300, default="")
     menu_name = models.CharField(max_length=100, default="")
-    rating = models.PositiveIntegerField(validators=[MinValueValidator(1), MaxValueValidator(5)], default=1)
+    rating = models.DecimalField(max_digits=2, decimal_places=1, validators=[MinValueValidator(1.0), MaxValueValidator(5.0)], default=1.0)
     created_at = models.DateTimeField(auto_now_add=True)
     like_count = models.PositiveIntegerField(default=0)
